intptr_t -> int32_t

+++ src/NetBox/FarConfiguration.cpp
@@ -232,9 +232,9 @@ intptr_t TFarConfiguration::GetConfirmationsSetting(HANDLE &Settings, const wcha
 }
 
 
-intptr_t TFarConfiguration::GetConfirmationsSettings() const
+int32_t TFarConfiguration::GetConfirmationsSettings() const
 {
-  intptr_t Result = 0;
+  int32_t Result = 0;
   FarSettingsCreate settings = {sizeof(FarSettingsCreate), FarGuid, INVALID_HANDLE_VALUE};
   HANDLE Settings = FFarPlugin->GetStartupInfo()->SettingsControl(INVALID_HANDLE_VALUE, SCTL_CREATE, 0, &settings) ? settings.Handle : 0;
   if (Settings)
@@ -269,7 +269,7 @@ void TFarConfiguration::CacheFarSettings()
   FFarConfirmations = GetConfirmationsSettings();
 }
 
-intptr_t TFarConfiguration::FarConfirmations() const
+int32_t TFarConfiguration::FarConfirmations() const
 {
   if (GetPlugin() && (GetCurrentThreadId() == GetPlugin()->GetFarThreadId()))
   {
